Fix bug and update doc
