+++ poulpy-cpu-avx/src/ntt4x30/tests.rs
@@ -208,6 +208,11 @@ fn test_convolution_direct() {
     test_convolution_accumulate_fused(&module, 50);
 }
 
+#[test]
+fn test_vmp_apply_dft_to_dft_digits_strided_bit_identical() {
+    poulpy_hal::test_suite::vmp::test_vmp_apply_dft_to_dft_digits_strided(&Module::<NTT4x30Avx>::new(64), 50);
+}
+
 cross_backend_test_suite! {
     mod word_compat,
     backend_ref =  poulpy_cpu_ref::NTT4x30Ref,
